Refactor script interpreter into focused submodules.
Split the 2.4k-line interpreter.rs into eval, push, rules, script_code, multisig, and tests modules without changing behavior.
